Cairo PDF backend asserts fix attempt part 1 (It's fucking shit because people hate me for bullshit fake reasons and won't help)
Instead of having the full fix or being able to use latest Cairo, everyone hates me for bullshit reasons that fall apart with logic or context and doesn't want to help me. So I'll either need to eventually make a part 2 or kang it from another fork if one is made.

This currently is just this commit ported over to the Firefox version of Cairo.
https://gitlab.freedesktop.org/cairo/cairo/-/merge_requests/266/diffs?commit_id=895b249b0ac6d2c45f47423f6529359991a66c3c

The full fix or part 2 is here, and my attempts to merge it so far have crashed and burned.
https://gitlab.freedesktop.org/cairo/cairo/-/commit/17ed347819d3928c0b050835ef3a17a891c124bc